% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_experiments.R
\name{choose_location}
\alias{choose_location}
\title{choose_location}
\usage{
choose_location(clustered_sf, n_locations)
}
\arguments{
\item{clustered_sf}{sf; An sf object with a column \code{cluster} designating
which cluster each polygon has been assigned to. Most like returned from
\code{finalize_clusters}.}

\item{n_locations}{integer; number of needed points.}
}
\value{
A matrix with columns \code{x, y} and nrows = \code{n_locations}. Values
represent the coordinates of the locations clicked on the map.
}
\description{
Choose locations of simulated expeirment
}
