% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_yields.R
\name{finalize_clusters}
\alias{finalize_clusters}
\title{Make final clusters}
\usage{
finalize_clusters(processed_data, cluster_number, mixing_parameter)
}
\arguments{
\item{processed_data}{sf; An sf object likely returned from \code{make_cluster_data}.
Object should include the features to be clustered and their associated polygon
geometries.}

\item{cluster_number}{integer; number of clusters to be used.}

\item{mixing_parameter}{numeric; Values to be passed as the alpha parameter
in \code{ClustGeo::hclustgeo} which controls the weighting between the data
and distance dissimilarity matrices. Value must be between 0 and 1.}
}
\value{
The sf object \code{processed_data} with a new column (cluster) giving
the cluster designation for each row.
}
\description{
Make final clusters
}
