% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{get_all_files}
\alias{get_all_files}
\title{get_all_files}
\usage{
get_all_files(path, files, file_ids, to_utm = TRUE)
}
\arguments{
\item{path}{string; directory where the files are stored}

\item{files}{string; names of files to be opened within the directory given by
\code{path}.}

\item{file_ids}{string; names to assign to new columns. Must be in the same
order as \code{files}.}

\item{to_utm}{logical; should crs be converted to UTM? Default it true.}
}
\value{
A named list of sf objects
}
\description{
Read one or more files
}
