% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{get_ashape_boundary_indices}
\alias{get_ashape_boundary_indices}
\title{get_ashape_boundary_indices}
\usage{
get_ashape_boundary_indices(ashape)
}
\arguments{
\item{ashape}{ashape; an object returned from alphahull::ashape.}
}
\value{
A vector of indices representing the indices of the xy matrix used to
create \code{ashape} that describe the shape's boundary.
}
\description{
Get the coordinate indices the descripe the boundary of an ashape object
}
\note{
adopted from https://rpubs.com/geospacedman/alphasimple and
https://babichmorrowc.github.io/post/2019-03-18-alpha-hull/
}
