% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_experiments.R
\name{get_block_origin}
\alias{get_block_origin}
\title{get_block_origin}
\usage{
get_block_origin(
  centroid,
  treatment_number,
  plot_length,
  plot_width,
  border_width
)
}
\arguments{
\item{centroid}{numeric vector; named vector of the xy point of the desired
block centroid. Must have the names \code{x} and \code{y}.}

\item{treatment_number}{integer; the number of treatments or plots that should
be included in each block.}

\item{plot_length}{numeric; desired plot length in meters.}

\item{plot_width}{numeric; desired plot width in meters.}

\item{border_width}{numeric; desired width between plots in meters.}
}
\value{
A named vector of xy coordinates describing the location of the bottom
left (south west) corner of one block.
}
\description{
Find the bottom left (south west) corner of a block
}
\note{
\code{border_width} is only used to to create borders between plots inside
the block. The block is draw with no external border.
}
