% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_experiments.R
\name{get_block_polygons}
\alias{get_block_polygons}
\title{get_block_polygons}
\usage{
get_block_polygons(
  treatment_number,
  plot_length,
  plot_width,
  border_width,
  crs,
  origin
)
}
\arguments{
\item{treatment_number}{integer; the number of treatments or plots that should
be included in each block.}

\item{plot_length}{numeric; desired plot length in meters.}

\item{plot_width}{numeric; desired plot width in meters.}

\item{border_width}{numeric; desired width between plots in meters.}

\item{crs}{crs; coordinate referense system to be applied to new polygons.}

\item{origin}{numeric vector; **optional** a named vector of the xy point
of the bottom left (south west) corner of the block. Must have the names
\code{x} and \code{y}. This should be provided if creating a connected
(traditional) experiment.}

\item{centroid}{numeric vector; **optional** a named vector of the xy point
of the desired block centroid. Must have the names \code{x} and \code{y}.
This should be provided if creating a disconnected experiment.}
}
\value{
A list of sf polygons with plot ids (1:treatment_number).
}
\description{
Create polygons of a block with plots
}
