% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{get_experiment_data}
\alias{get_experiment_data}
\title{Subset simulation data to experimental polygons}
\usage{
get_experiment_data(experiment_list, simulation_data)
}
\arguments{
\item{experiment_list}{named list; named list of sf objects returned
from \code{make_experiment} describing experiment polygons. Must be named with
the strings that should be used as ID's.}

\item{simulation_data}{sf; A named list of sf objects returned from get_test_data.
This is the data that will be used in simulation experiments.}
}
\value{
An sf object of the simulation_data point observatoins, subset to only
though intersecting with the polygons in experiment_list. Each point is
identified by id (name of experiment from experiment_list), block, and plot.
If experiments overlap, the overlapping points will be once for each experiment.
}
\description{
Subset simulation data to experimental polygons
}
