% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_experiments.R
\name{get_experiment_origins}
\alias{get_experiment_origins}
\title{Find the bottom left (south west) corner of each block in a connected experiment}
\usage{
get_experiment_origins(
  centroid,
  treatment_number,
  plot_length,
  plot_width,
  border_width,
  block_rows,
  block_cols
)
}
\arguments{
\item{centroid}{numeric vector; named vector of the xy point of the center of
the whole experiment. Must have the names \code{x} and \code{y}.}

\item{treatment_number}{integer; the number of treatments or plots that should
be included in each block.}

\item{plot_length}{numeric; desired plot length in meters.}

\item{plot_width}{numeric; desired plot width in meters.}

\item{border_width}{numeric; desired width between plots in meters.}

\item{block_rows}{integer; number of rows the blocks should be arranged into.
\code{block_rows * block_cols} equals total number of blocks.}

\item{block_cols}{integer; number of cols the blocks should be arranged into.
\code{block_rows * block_cols} equals total number of blocks.}
}
\value{
A matrix of the xy origins (bottom left/south west corner) of each
block in an experiment. The matrix columns are named x and y.
}
\description{
Find the bottom left (south west) corner of each block in a connected experiment
}
