% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{get_field_boundary}
\alias{get_field_boundary}
\title{Create a field boundary from points in an sf object}
\usage{
get_field_boundary(field, alpha)
}
\arguments{
\item{field}{sf; spatial object with points}

\item{alpha}{numeric; alpha value to be passed to \code{alphahull::ashape}.}
}
\value{
A polygon of the field boundary
}
\description{
Create a field boundary from points in an sf object
}
