% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{get_field_grid}
\alias{get_field_grid}
\title{get_field_grid}
\usage{
get_field_grid(field, alpha, harvest_width, passes_to_clip, cellsize_scaler)
}
\arguments{
\item{field}{sf; an sf object of point geometries.}

\item{alpha}{numeric; parameter that controls the level of simplication in the
field boundary. Larger numbers are more simple and follow the data points less closely.
This parameter is passed to \code{alphahull::ashape}.}

\item{harvest_width}{numeric; width of harvest header, in meters.}

\item{passes_to_clip}{integer; the number of harvest passes to clip when creating
field buffer.}

\item{cellsize_scaler}{numeric; controls the size of the grid cells. The value
of \code{combind_width * cellsize_scaler} is passed to \code{sf::st_make_grid}.}
}
\value{
An sfc_POLYGON object that contains a hexagonal grid of cell size
\code{harvest_width * cellsize_scaler}, clipped to the size of the field buffer.
The field buffer is the detected field boundardy, with its simplifiction controlled
by \code{alpha}, minus \code{harvest_width * passes_to_clip}.
}
\description{
Make polygon grid of field
}
\note{
TODO maybe add ... to give more flexibility in passing arguments?
TODO smarter way to choose field that is used to create the grid.
}
