% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_experiments.R
\name{get_plot_boundary}
\alias{get_plot_boundary}
\title{get_plot_boundary}
\usage{
get_plot_boundary(plot_length, plot_width, x0, y0)
}
\arguments{
\item{plot_length}{numeric; desired plot length in meters.}

\item{plot_width}{numeric; desired plot width in meters.}

\item{x0}{numeric; the x value of the bottom left (south west) corner of the
plot.}

\item{y0}{numeric; the y value of the bottom left (south west) corner of the
plot.}
}
\value{
A matrix of the five points describing the polygon of a singe plot. There
are five points because polygons must be closed with the first and last value
repeated.
}
\description{
Create a matrix of the points defining a single plot polygon
}
