% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{get_test_data}
\alias{get_test_data}
\title{Get a yields to use in simulation}
\usage{
get_test_data(path, file, var_of_interest)
}
\arguments{
\item{var_of_interest}{string; can be multiple}
}
\value{
An sf data frame of the original point yields, converted to UTM
}
\description{
Get a yields to use in simulation
}
