% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{get_utm}
\alias{get_utm}
\title{get_utm}
\usage{
get_utm(long)
}
\arguments{
\item{long}{numeric; longitude. Longitude should be given as degrees East.
Examples: San Fransico, CA is -122, Sydney Australia is 151.}
}
\value{
UTM zone
}
\description{
Find the appropriate UTM based on longitude
}
\note{
based on https://stackoverflow.com/questions/9186496/determining-utm-zone-to-convert-from-longitude-latitude
}
