% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{input_checker}
\alias{input_checker}
\title{input_checker}
\usage{
input_checker(config, which_function)
}
\arguments{
\item{config}{list; a named list containing all the needed inputs.}

\item{which_function}{string; string indicating which requirement list should
be used. The current options are "make_cluster_data" or "make_experiment".}
}
\value{
If there are any missing names or mismatched types between the requirement
list and the config, an error will be returned. Otherwise, nothing is returned.
}
\description{
Check configuration file
}
