% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MAKE_CLUSTER_REQUIRED}
\alias{MAKE_CLUSTER_REQUIRED}
\title{Required names and types for all inputs used in make_cluster_data}
\format{
A list with 9 elements. Each list element is a named vector with elements:
\describe{
  \item{name}{the required inputs name}
  \item{type}{the type the inputs should be (character, numeric, etc.)}
}
}
\usage{
MAKE_CLUSTER_REQUIRED
}
\description{
Required names and types for all inputs used in make_cluster_data
}
\keyword{datasets}
