% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_yields.R
\name{make_dendrogram_plot}
\alias{make_dendrogram_plot}
\title{make_dendrogram_plot}
\usage{
make_dendrogram_plot(plot_data, tree, cluster_number)
}
\arguments{
\item{plot_data}{sf; a copy of the \code{processed_data} passed to
\explore{explore_dendrogram}. \code{processed_data} is an sf object, likely
returned from \code{make_cluster_data}. Object should include the features to
be clustered and their associated polygon geometries.}

\item{tree}{hclust; an hclust object returned from \code{ClustGeo::hclustgeo}.}

\item{cluster_number}{integer; number of clusters to be used.}
}
\value{
A two panel plot with the first panel being the field map colored by
cluster and the second panel being a dedrogram with colored rectangles showing
the clusters. The colors in the two plots should match.
}
\description{
Create active binding to plot code in a split screen device
}
