% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_experiments.R
\name{make_disconnect_exp}
\alias{make_disconnect_exp}
\title{make_disconnect_exp}
\usage{
make_disconnect_exp(
  experiment_type,
  clustered_sf,
  n_locations,
  treatment_number,
  plot_length,
  plot_width,
  border_width,
  crs,
  block_cols = NULL,
  block_rows = NULL,
  rotation_angle = NULL
)
}
\arguments{
\item{clustered_sf}{sf; An sf object with a column \code{cluster} designating
which cluster each polygon has been assigned to. Most like returned from
\code{finalize_clusters}.}

\item{n_locations}{integer; number of needed points.}

\item{treatment_number}{integer; the number of treatments or plots that should
be included in each block.}

\item{plot_length}{numeric; desired plot length in meters.}

\item{plot_width}{numeric; desired plot width in meters.}

\item{border_width}{numeric; desired width between plots in meters.}

\item{crs}{crs; coordinate referense system to be applied to new polygons.}

\item{rotaion_anlge}{integer; **optional** angle by which the original experiment should
be rotated. Default is NULL, which will result in plots oriented North-South.}
}
\value{
An sf object containing the polygons for each plot in an experiment.
}
\description{
Make a disconnected experiment
}
