% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MAKE_EXPERIMENT_REQUIRED}
\alias{MAKE_EXPERIMENT_REQUIRED}
\title{Required names and types for all inputs used in make_experiment}
\format{
A list with 5 elements. Each list element is a named vector with elements:
\describe{
  \item{name}{the required inputs name}
  \item{type}{the type the inputs should be (character, numeric, etc.)}
}
}
\usage{
MAKE_EXPERIMENT_REQUIRED
}
\description{
Required names and types for all inputs used in make_experiment
}
\keyword{datasets}
