% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{no_side_effect_warning}
\alias{no_side_effect_warning}
\title{Stop execution if nothing is going to be returned or saved}
\usage{
no_side_effect_warning(plot, output_path)
}
\arguments{
\item{plot}{logical; should plot be returned in the current graphics
device (likely in RStudio).}

\item{output_path}{string; path where plot should be saved.
Default is NULL. If path is provided, a PDF with the name "dendrogram_k_clusters"
will be created, where k is the integer passed to cluster_number.}
}
\value{
If plot = FALSE and output_path = NULL, an error is returned. Otherwise,
nothing is returned.
}
\description{
Stop execution if nothing is going to be returned or saved
}
