% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_experiments.R
\name{plot_experiment}
\alias{plot_experiment}
\title{Plot experiment polygons}
\usage{
plot_experiment(experiment)
}
\arguments{
\item{experiment}{sf; an sf object containing the polygons for each
plot and/or blocks.}
}
\value{
Addes the plotted polygons to an existing map of the experiment with
labels containing the block numbers.
}
\description{
Plot experiment polygons
}
