% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_experiments.R
\name{rotate_experiment}
\alias{rotate_experiment}
\title{Rotate either connected or disconnected experiment by given angle}
\usage{
rotate_experiment(original_experiment, rotation_angle, experiment_type)
}
\arguments{
\item{original_experiment}{sf; an sf object containing the polygons for each
plot and/or blocks. The rotation will occur around the centroid of \code{original_experiment}.
Therefore, if each block should be rotated individually, they need to be passed
one at a time.}

\item{experiment_type}{string; which experiment type should be created? Current
options are "connected", which creates as single traditional experiment and
expects the user to select only one centroid, or "disconnected" which expects
the user to select the centroid of each block independently.}

\item{rotaion_anlge}{integer; angle by which the original experiment should
be rotated.}
}
\value{
The rotated sf object. Will have the same dimentions as \code{original_experiment}
but with the POLYGON geometries rotated.
}
\description{
Rotate either connected or disconnected experiment by given angle
}
