% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_experiments.R
\name{rotate_polygon}
\alias{rotate_polygon}
\title{Merge experiment or block polygon and rotate by a give angle}
\usage{
rotate_polygon(original_experiment, rotation_angle)
}
\arguments{
\item{original_experiment}{sf; an sf object containing the polygons for each
plot and/or blocks. The rotation will occur around the centroid of \code{original_experiment}.
Therefore, if each block should be rotated individually, they need to be passed
one at a time.}

\item{rotaion_anlge}{integer; angle by which the original experiment should
be rotated.}
}
\value{
The rotated sf object. Will have the same dimentions as \code{original_experiment}
but with the POLYGON geometries rotated.
}
\description{
Merge experiment or block polygon and rotate by a give angle
}
\note{
Taken from https://stackoverflow.com/questions/51282724/creating-a-regular-polygon-grid-over-a-spatial-extent-rotated-by-a-given-angle
}
