% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_experiments.R
\name{rotate_single}
\alias{rotate_single}
\title{Rotation around a single point}
\usage{
rotate_single(a)
}
\arguments{
\item{a}{numeric}
}
\value{
rotated point
}
\description{
Rotation around a single point
}
\note{
taken from https://r-spatial.github.io/sf/articles/sf3.html
}
