% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{save_fig}
\alias{save_fig}
\title{save_fig}
\usage{
save_fig(path, name, plot_call)
}
\arguments{
\item{path}{string; path to the directory where the figure should be saved.
If the directory does not already exist, it will be created. If there is
already a figure in the directory with the same name, it will be overwritten.}

\item{name}{string; name that should be given to the figure. Do not include ".pdf".}

\item{plot_call}{plot; either the code to generate a plot (ex: \code{plot(1:10)})
or an active binding to code that generates a plot.}
}
\value{
A saved pdf of the figure created by \code{plot_call} in the directory
named by \code{path}.
}
\description{
Save a figure as a pdf
}
