% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{update_field_crs}
\alias{update_field_crs}
\title{update_field_crs}
\usage{
update_field_crs(field)
}
\arguments{
\item{field}{sf; an sf object. The longitude of the bbox is used to find the
correct UTM zone.}
}
\value{
An updated sf object with a UTM CRS added.
}
\description{
Update field CRS to UTM zone
}
\note{
This is too simple for wide use.
}
