% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_yields.R
\name{explore_dendrogram}
\alias{explore_dendrogram}
\title{explore_dedrogram}
\usage{
explore_dendrogram(processed_data, cluster_number)
}
\arguments{
\item{processed_data}{sf; An sf object likely returned from \code{make_cluster_data}.
Object should include the features to be clustered and their associated polygon
geometries.}

\item{cluster_number}{integer; number of clusters to be used.}
}
\value{
A two panel plot with the first panel being the field map colored by
cluster and the second panel being a dedrogram with colored rectangles showing
the clusters. The colors in the two plots should match.
}
\description{
Make dendrogram and single map
}
